Remove flutter binary
