+++ lib/commands/flutter.dart
@@ -3,6 +3,9 @@ import 'package:fvm/constants.dart';
 import 'package:fvm/flutter/flutter_tools.dart';
 import 'package:fvm/utils/helpers.dart';
 import 'package:args/args.dart';
+import 'package:fvm/utils/print.dart';
+import 'package:fvm/utils/project_config.dart';
+import 'package:process_run/which.dart';
 
 /// Proxies Flutter Commands
 class FlutterCommand extends Command {
@@ -21,7 +24,18 @@ class FlutterCommand extends Command {
 
   @override
   Future<void> run() async {
-    final flutterExec = getFlutterSdkExec();
+    var flutterExec = getFlutterSdkExec();
+    if (flutterExec == '') {
+      final globalFlutter = await which('flutter');
+      if (globalFlutter == '') {
+        throw Exception('FVM: Flutter not found in path');
+      }
+      flutterExec = globalFlutter;
+      PrettyPrint.info('FVM: Global Flutter\nPath: $flutterExec\n');
+    } else {
+      PrettyPrint.info(
+          'FVM: Local Flutter ${getConfigFlutterVersion()}\nPath:($flutterExec)\n');
+    }
 
     await flutterCmd(flutterExec, argResults.arguments,
         workingDirectory: kWorkingDirectory.path);
