+++ pubspec.yaml
@@ -8,7 +8,6 @@ environment:
 
 executables:
   fvm: main
-  flutter: flutter
 
 dependencies:
   args: ^1.5.3
